+++ src/index.ts
@@ -13,7 +13,7 @@ export async function scheduleShutdown(
 ): Promise<Date> {
     const minutes = parseAndValidateMinutes(duration);
     const start = Date.now();
-    await getDriver().scheduleShutdown(minutes, verbose);
+    await getDriver(process.platform).scheduleShutdown(minutes, verbose);
     return new Date(start + minutes * 60 * 1000);
 }
 
@@ -29,7 +29,7 @@ export async function scheduleRestart(
 ): Promise<Date> {
     const minutes = parseAndValidateMinutes(duration);
     const start = Date.now();
-    await getDriver().scheduleRestart(minutes, verbose);
+    await getDriver(process.platform).scheduleRestart(minutes, verbose);
     return new Date(start + minutes * 60 * 1000);
 }
 
@@ -38,5 +38,5 @@ export async function scheduleRestart(
  * @param verbose - If set to true, the internal commands and their outputs will be logged
  */
 export async function cancelShutdown(verbose: boolean = false): Promise<void> {
-    await getDriver().cancelShutdown(verbose);
+    await getDriver(process.platform).cancelShutdown(verbose);
 }