+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 -   This file
 -   Apply Prettier to .md files
 -   Use `files` property in package.json instead of .npmignore
+-   Add more tests and test coverage
 
 ## [1.1.2] - 2018-12-04